refactor: Removed the navBar from the homescreen into a component. Passed state in order to maintain backDrop handlers.
